Correction using an EPI and a T1w

+++ dipy/align/__init__.py
@@ -1,10 +1,10 @@
 import numpy as np
-floating  = np.float32
+floating  = np.float64
 
 class Bunch(object):
     def __init__(self, **kwds):
         r"""A 'bunch' of values (a replacement of Enum)
-        
+
         This is a temporary replacement of Enum, which is not available
         on all versions of Python 2
         """
